+++ ReleaseNotes.md
@@ -5,7 +5,8 @@
 ## Changes
 
 ## Bugfixes
-
+- Fixed Secondary battery and AA defense armaments not having localization.
+- Fixed Blast penetration parameter being shown for AP and SAP.
 ___
 To download the program for the first time, you need to download only the Setup.exe file. The other files are used by the application for automatic updates.